+++ src/helpers/execScript.ts
@@ -1,9 +1,10 @@
-import { exec, spawn } from 'child_process';
+import { exec } from 'child_process';
+import { spawn } from 'cross-spawn';
 
 export const execScript = async <T>(command: string, args: string[] = [], shouldSpawn: boolean = false): Promise<T> => {
   return new Promise<T>((resolve, reject) => {
     if (shouldSpawn) {
-      const execution = spawn(command, [...args], { shell: process.platform === "win32" });
+      const execution = spawn(command, [...args], { stdio: 'inherit' });
 
       execution.stdout.on('data', (data) => {
         console.log(`${data}`);